+++ lib/belt.rb
@@ -1,3 +1,8 @@
+# frozen_string_literal: true
+
+require_relative "belt/version"
+require_relative "belt/parameters"
+require_relative "belt/application_controller"
+
 module Belt
-  VERSION = "0.0.1"
 end